Fix integer value for `GL_MAX_UNIFORM_BLOCK_SIZE` overflowing

+++ drivers/gles3/shader_gles3.h
@@ -154,7 +154,7 @@ private:
 	static bool shader_cache_save_debug;
 	bool shader_cache_dir_valid = false;
 
-	GLint max_image_units = 0;
+	int64_t max_image_units = 0;
 
 	enum StageType {
 		STAGE_TYPE_VERTEX,
